+++ bundles/framework/bundle/mapmodule-plugin/plugin/wmslayer/WmsLayerPlugin.js
@@ -178,9 +178,6 @@ Oskari.clazz.define('Oskari.mapframework.mapmodule.WmsLayerPlugin',
             }
 
         },
-        addMapLayerToMap: function (layer, keepLayerOnTop, isBaseMap) {
-            this.addMapLayerToMap(layer, keepLayerOnTop, isBaseMap);
-        },
         /**
          * Adds a single WMS layer to this map
          *