+++ bundles/framework/bundle/mapmyplaces/plugin/MyPlacesLayerPlugin.js
@@ -267,9 +267,6 @@ function(config) {
         }
 
     },
-    addMapLayerToMap: function(layer, keepLayerOnTop, isBaseMap) {
-        this.addMapLayerToMap(layer, keepLayerOnTop, isBaseMap);
-    },
     /**
      * Adds a single MyPlaces layer to this map
      *