+++ sources/framework/domain/AbstractLayer.js
@@ -601,7 +601,22 @@ Oskari.clazz.define('Oskari.mapframework.domain.AbstractLayer',
          * adds style to layer
          */
         addStyle: function (style) {
-            this.getStyles().push(style);
+            if(!style || !style.getName || typeof style.getName !== 'function') {
+                // invalid style
+                return;
+            }
+            var foundExisting = false;
+            for(var i = 0; i < this.getStyles().length; ++i) {
+                var curStyle = this.getStyles()[i];
+                if(curStyle.getName() == style.getName()) {
+                    foundExisting = true;
+                    break;
+                }
+            }
+            if(!foundExisting) {
+                // only add if one with same name isn't added yet
+                this.getStyles().push(style);
+            }
         },
         /**
          * @method getStyles
@@ -623,10 +638,6 @@ Oskari.clazz.define('Oskari.mapframework.domain.AbstractLayer',
         selectStyle: function (styleName, preventRecursion) {
             var me = this;
 
-            // Layer doesn't have styles
-            if (me.getStyles().length === 0) {
-                this.addStyle(this._createEmptyStyle());
-            }
             for (var i = 0; i < me.getStyles().length; i++) {
                 var style = me.getStyles()[i];
                 if (style.getName() === styleName) {
@@ -639,16 +650,9 @@ Oskari.clazz.define('Oskari.mapframework.domain.AbstractLayer',
             }
 
             // didn't match anything select the first one
-            if(!preventRecursion && !me._currentStyle) {
-                if(styleName === '') {
-                    // FIXME: Fix layer styles array so it doesn't contain invalid styles so this isn't needed
-                    // tried to select empty style
-                    this._currentStyle = this._createEmptyStyle();
-                }
-                else {
-                    // tried to select non-existing style
-                    this.selectStyle(me.getStyles()[0].getName(), true);
-                }
+            if(!me._currentStyle) {
+                // Style not found, use an empty one!
+                this._currentStyle = this._createEmptyStyle();
             }
         },
         /**