+++ ReleaseNotes.md
@@ -6,6 +6,20 @@
 
 Removed handling of `AfterMapLayerAddEvent` from the layer plugins for it is the mapmodule who handles it now. It calls `addMapLayerToMap` function for each of its registered layer plugins and assures the marker layer always appear on top of the map.
 
+### mapwfs2
+
+WfsLayerPlugin now assumes config values hostname and port based on document.location if not configured and contextPath also defaults to '/transport' if not configured.
+
+### core/map layer service/AbstractLayer
+
+Maplayer JSON parsing changed a bit:
+
+* Legendimage is now parsed for all layer types
+
+* AbstractLayer.addStyle() now checks that a style with the same name isn't added yet.
+
+* Formats parsing has been moved out from styles parsing and in to wmslayer specific parsing as they are not related operations
+
 ## 1.17
 
 ### **Breaking changes**